Add flags option to table loading
When loading a verity table the device *must* be a readonly. The dm
driver requires the flags for the DM_TABLE_LOAD_CMD ioctl to include
DM_READONLY.
Add an optiona argument to `core::dm::table_load` that enables any
user to pass DmFlags::DM_READONLY to the DM_TABLE_LOAD_CMD operation.